Create multitenancy.md
If the "What it is" section looks too verbose, the last 1 or even 2 paragraphs of it can be deleted, although I have re-read everything carefully and do believe that they provide important supplementary information that is not redundant.

I confirm that this is completely my own work, although in the world of term definitions you can always find fragments similar to each other even if they were developed independently.

Signed-off-by: Sergey Borovoy <48699431+sergeyborovoy@users.noreply.github.com>