[ES] Update term reference links
Since not all the glossary terms were available at the same time,
contributors deferred from referencing terms in Spanish. Now, that
we have the majority of terms translated we need to double check
we are using the correct references.

For the record this can be easily spotte with the following grep
expression:

  grep -PHnro "\[.*?\].*?\(/.*?\)" content/es | grep -v '(/es/'

It will capture and display all the reference links that are not
using the Spanish terms.

PS: I omitted SaaS since it got deprecated and its removal was
already done in the main branch.

Signed-off-by: Jonnatan Jossemar Cordero <jossemargt@gmail.com>