Create host.md
This has been a long time coming after I initially submitted this idea years ago!

As best as I can tell, this is similar to the "service" / "microservice" relationship, so I've tried to model this doc after the "service" one. I've added a bit of context about "host" that isn't present in "nodes", but maybe it's unnecessary? Feedback definitely appreciated. 

Signed-off-by: Jason Rhodes <jason.matthew.rhodes@gmail.com>